requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
coverage==7.10.6
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.119.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.10
pydantic==2.12.0
pydantic_core==2.41.1
pygount==3.1.0
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
ruff==0.14.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
starlette==0.48.0
tree-sitter==0.25.2
tree_sitter_language_pack==0.10.0
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
pypdf>=3.0.0
python-docx==1.2.0
google-generativeai>=0.8.0
